+++ crates/schematic/src/schema/renderers/mod.rs
@@ -1,20 +1,20 @@
-#[cfg(feature = "json_schema")]
+#[cfg(feature = "renderer_json_schema")]
 pub mod json_schema;
 
-#[cfg(all(feature = "template", feature = "json"))]
+#[cfg(all(feature = "renderer_template", feature = "json"))]
 pub mod json_template;
 
-#[cfg(all(feature = "template", feature = "json"))]
+#[cfg(all(feature = "renderer_template", feature = "json"))]
 pub mod jsonc_template;
 
-#[cfg(feature = "template")]
+#[cfg(feature = "renderer_template")]
 pub mod template;
 
-#[cfg(all(feature = "template", feature = "toml"))]
+#[cfg(all(feature = "renderer_template", feature = "toml"))]
 pub mod toml_template;
 
-#[cfg(feature = "typescript")]
+#[cfg(feature = "renderer_typescript")]
 pub mod typescript;
 
-#[cfg(all(feature = "template", feature = "yaml"))]
+#[cfg(all(feature = "renderer_template", feature = "yaml"))]
 pub mod yaml_template;