+++ crates/schematic/tests/defaults_test.rs
@@ -141,7 +141,7 @@ fn handles_nested_defaults() {
     assert!(!result.config.nested.boolean);
 }
 
-#[cfg(feature = "json_schema")]
+#[cfg(feature = "renderer_json_schema")]
 #[test]
 fn generates_json_schema() {
     use starbase_sandbox::{assert_snapshot, create_empty_sandbox};
@@ -162,7 +162,7 @@ fn generates_json_schema() {
     assert_snapshot!(std::fs::read_to_string(file).unwrap());
 }
 
-#[cfg(feature = "typescript")]
+#[cfg(feature = "renderer_typescript")]
 #[test]
 fn generates_typescript() {
     use starbase_sandbox::{assert_snapshot, create_empty_sandbox};