+++ crates/schematic/tests/env_test.rs
@@ -208,7 +208,7 @@ fn loads_from_prefixed() {
     assert_eq!(result.config.list2, vec![1, 2, 3]);
 }
 
-#[cfg(feature = "json_schema")]
+#[cfg(feature = "renderer_json_schema")]
 #[test]
 fn generates_json_schema() {
     use starbase_sandbox::{assert_snapshot, create_empty_sandbox};
@@ -226,7 +226,7 @@ fn generates_json_schema() {
     assert_snapshot!(std::fs::read_to_string(file).unwrap());
 }
 
-#[cfg(feature = "typescript")]
+#[cfg(feature = "renderer_typescript")]
 #[test]
 fn generates_typescript() {
     use starbase_sandbox::{assert_snapshot, create_empty_sandbox};