+++ crates/schematic/tests/generator_test.rs
@@ -147,7 +147,7 @@ fn create_template_options() -> TemplateOptions {
     }
 }
 
-#[cfg(feature = "json_schema")]
+#[cfg(feature = "renderer_json_schema")]
 mod json_schema {
     use super::*;
     use schematic::schema::json_schema::*;
@@ -231,7 +231,7 @@ mod template_yaml {
     }
 }
 
-#[cfg(feature = "typescript")]
+#[cfg(feature = "renderer_typescript")]
 mod typescript {
     use super::*;
     use schematic::schema::typescript::*;